+++ tests/main_test.py
@@ -27,6 +27,7 @@ def test_dry_run(workspace_name: str, beaker_cluster_name: str, run_name: str):
             workspace_name,
             "--cluster",
             beaker_cluster_name,
+            "--yes",
             "--",
             "python",
             "-c",