+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added additional `--dry-run` output.
+- Added `--yes` flag for skipping confirmation prompts.
+
+### Changed
+
+- Public workspaces are now allowed.
+- Workspaces with multiple contributors are now allowed, but you'll be asked to confirm
+  that it's okay.
+
 ## [v0.1.0](https://github.com/allenai/beaker-gantry/releases/tag/v0.1.0) - 2022-05-13
 
 ### Added